+++ kawow/__init__.py
@@ -21,7 +21,7 @@ R. Naef, W.E. Acree Jr., Liquids 4(1):231-260, 2024.
 DOI: 10.3390/liquids4010011
 """
 
-from .model import PartitionCalculator, MQGPartitionCalculator, EnsemblePartitionCalculator, PFASGroupsPartitionCalculator, fit, fit_mqg, run_models
+from .model import PartitionCalculator, MQGPartitionCalculator, EnsemblePartitionCalculator, PFASGroupsPartitionCalculator, PFASGroupsRFPartitionCalculator, PFASGroupsXGBPartitionCalculator, PFASGroupsNNPartitionCalculator, fit, fit_mqg, run_models
 from .smarts_model import NaefAcreePartitionCalculator, NaefAcreeCrippenMixedPartitionCalculator
 from .io import parse_input
 from .features import compute_features
@@ -33,6 +33,9 @@ __all__ = [
     "MQGPartitionCalculator",
     "EnsemblePartitionCalculator",
     "PFASGroupsPartitionCalculator",
+    "PFASGroupsRFPartitionCalculator",
+    "PFASGroupsXGBPartitionCalculator",
+    "PFASGroupsNNPartitionCalculator",
     "NaefAcreePartitionCalculator",
     "NaefAcreeCrippenMixedPartitionCalculator",
     "fit",